Users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-mail = "^0.10.0"
flask-bcrypt = "^1.0.1"
flask-jwt-extended = "^4.7.1"
python-dotenv = "^1.0.1"
itsdangerous = "^2.2.0"
flask-script = "^2.0.6"
pytest = "^8.3.4"
pytest-flask = "^1.3.0"
flask-marshmallow = "^1.3.0"
flask-migrate = "^4.1.0"
flask-wtf = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
